Move and rename passwords.go from util to hash.go to auth package
